+++ test/cldr_test.exs
@@ -297,7 +297,7 @@ defmodule Cldr.Test do
   end
 
   test "Cldr.Chars.to_string for a language_tag" do
-    {:ok, locale} = Cldr.validate_locale("en-US-u-cu-AUD-nu-thai", MyApp.Cldr)
-    assert Cldr.to_string(locale) == "en-Latn-US-u-cu-AUD-nu-thai"
+    {:ok, locale} = Cldr.validate_locale("en-US-u-nu-thai-cu-AUD", MyApp.Cldr)
+    assert Cldr.to_string(locale) == "en-Latn-US-u-cu-aud-nu-thai"
   end
 end